Seasonal Aggregation copmlete
While the buttons aren't dynamic, you can aggregate based on the current hard-coded options available. It will work similarly to the aggregate page, but now only for the time frame of the given season. On top of that, it will only display stats for skiing. Also, an icon for seasonal aggregation was created last commit